+++ PKHeX.Core/PKM/PK2.cs
@@ -165,7 +165,7 @@ namespace PKHeX.Core
             };
             pk7.Language = GuessedLanguage(PKMConverter.Language);
             pk7.Nickname = PKX.GetSpeciesNameGeneration(pk7.Species, pk7.Language, pk7.Format);
-            if (otname[0] == StringConverter.G1TradeOTCode) // Ingame Trade
+            if (otname[0] == StringConverter12.G1TradeOTCode) // Ingame Trade
                 pk7.OT_Name = Encounters1.TradeOTG1[pk7.Language];
             pk7.OT_Friendship = pk7.HT_Friendship = PersonalTable.SM[Species].BaseFriendship;
 
@@ -194,10 +194,10 @@ namespace PKHeX.Core
             {
                 pk7.IsNicknamed = true;
                 pk7.Nickname = Korean ? Nickname
-                    : StringConverter.GetG1ConvertedString(nick, Japanese);
+                    : StringConverter12.GetG1ConvertedString(nick, Japanese);
             }
             pk7.OT_Name = Korean ? OT_Name
-                : StringConverter.GetG1ConvertedString(otname, Japanese);
+                : StringConverter12.GetG1ConvertedString(otname, Japanese);
             pk7.OT_Gender = OT_Gender; // Crystal
 
             pk7.TradeMemory(Bank: true); // oh no, memories on gen7 pkm
