+++ PKHeX.Core/Saves/Substructures/Mail/Mail4.cs
@@ -55,7 +55,7 @@ namespace PKHeX.Core
         public byte AuthorLanguage { get => Data[5]; set => Data[5] = value; }
         public byte AuthorVersion { get => Data[6]; set => Data[6] = value; }
         public override int MailType { get => Data[7]; set => Data[7] = (byte)value; }
-        public override string AuthorName { get => StringConverter.GetString4(Data, 8, 0x10); set => StringConverter.SetString4(value, 7, 8, 0xFFFF).CopyTo(Data, 8); }
+        public override string AuthorName { get => StringConverter4.GetString4(Data, 8, 0x10); set => StringConverter4.SetString4(value, 7, 8, 0xFFFF).CopyTo(Data, 8); }
         public int GetAppearPKM(int index) => BitConverter.ToUInt16(Data, 0x1C - (index * 2));
         public void SetAppearPKM(int index, int value) => BitConverter.GetBytes((ushort)(value == 0 ? 0xFFFF : value)).CopyTo(Data, 0x1C - (index * 2));
         public override ushort GetMessage(int index1, int index2) => BitConverter.ToUInt16(Data, 0x20 + (((index1 * 4) + index2) * 2));
